requirements/base.txt
amqp==5.2.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
attrs==24.2.0
billiard==4.2.0
binaryornot==0.4.4
bleach==6.1.0
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cookiecutter==2.6.0
crispy-bootstrap5==2024.2
cron-descriptor==1.4.5
cryptography==43.0.1
Django==5.0.9
django-allauth==64.2.1
django-appconf==1.0.6
django-celery-beat==2.7.0
django-compressor==4.5.1
django-cors-headers==4.4.0
django-crispy-forms==2.3
django-environ==0.11.2
django-extensions==3.2.3
django-filter==24.2
django-jazzmin==3.0.0
django-model-utils==5.0.0
django-redis==5.4.0
django-summernote==0.8.20.0
django-timezone-field==7.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
Faker==28.4.1
fido2==1.1.3
flower==2.0.1
hiredis==3.0.0
humanize==4.10.0
idna==3.10
inflection==0.5.1
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kombu==5.4.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
pillow==10.4.0
prometheus_client==0.20.0
prompt_toolkit==3.0.47
pycparser==2.22
Pygments==2.18.0
PyJWT==2.9.0
pypng==0.20220715.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-slugify==8.0.4
pytz==2024.2
PyYAML==6.0.2
qrcode==7.4.2
rcssmin==1.1.2
redis==5.0.8
referencing==0.35.1
requests==2.32.3
rich==13.8.1
rjsmin==1.2.2
rpds-py==0.20.0
six==1.16.0
sqlparse==0.5.1
text-unidecode==1.3
tornado==6.4.1
types-python-dateutil==2.9.0.20240906
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==3.0.4
whitenoise==6.7.0
